Fixed thread problems

+++ Auxiliary/Auxiliary.cs
@@ -9,7 +9,7 @@ namespace CupheadArchipelago.Auxiliary {
     public class Aux {
         public static string BlankIsNull(string str) => str.Length>0?str:null;
 
-        public static bool IntAsBool(int i) => i != 0;
+        public static bool IntAsBool(long i) => i != 0;
 
         public static string CollectionToString(IEnumerable collection) {
             bool first = true;
